Refactorizacion para dinero

+++ src/fiuba/algo3/tp2/modelo/encasillables/Quini6.java
@@ -6,7 +6,7 @@ import fiuba.algo3.tp2.modelo.Jugador;
 
 public class Quini6 implements Encasillable{
 
-	static final Dinero monto_a_incrementar = new Dinero((double) 50000);
+	static final Dinero monto_a_incrementar = new Dinero(50000);
 	
 	@Override
 	public void ejecutarEfecto(Jugador jugador) {
